seperate EditTagsModal into its own component
